Refactor grammar tests into tests/grammar/ with MEDSDataset + pytest options
Addresses the design-level review comments on PR #106:

1. **CLI warning for `do_sample=false`** (`__main__.py`): fires a warning at
   generate-trajectories start if `inference.do_sample=false`. Greedy is an
   anti-pattern for real trajectory generation (collapses N_trajectories
   diversity); only legitimate in correctness regression tests like our
   grammar fixture, which deliberately trips the warning.

2. **tests/grammar/ subfolder**: moves all grammar-related files into one
   package:
   - `_grammar.py`  — grammar definitions (PROGRAMS, SEP, GrammarFSM,
     sample_sequence, build_training_batch_codes, mock_batch), extracted
     from the former `test_pattern_generation.py`.
   - `_meds.py`    — MEDS round-trip, now using
     `meds_testing_helpers.dataset.MEDSDataset.write()` instead of hand-
     rolled parquet writes (drops ~60 LOC of manual schema definitions,
     `dataset.json` JSON literal, and per-split parquet writes; relies on
     the canonical `meds.DatasetMetadataSchema` / `MEDSDataset.PL_LABEL_SCHEMA`
     so we don't drift on a MEDS-spec bump).
   - `conftest.py` — pytest options (`--grammar-pretrain-epochs=...` etc.)
     replacing the old GRAMMAR_* module-level constants, plus the grammar
     fixtures that consume those options. Fixtures are auto-discovered by
     pytest under `tests/grammar/`; no cross-package import plumbing.
   - `test_in_process.py` — the in-process tests moved from the old
     `test_pattern_generation.py`.
   - `test_cli.py` — the CLI tests moved from the old
     `test_pattern_generation_cli.py`, with hardcoded constants replaced
     by fixture-injected values (e.g. `grammar_n_trajectories` instead of
     `GRAMMAR_N_TRAJECTORIES`).

3. **Use meds-testing-helpers' MEDSDataset for dataset construction**
   (folded into #2 above). See comment response on PR for why this is
   worth ~60 LOC of simplification plus MEDS-spec drift protection.

4. **Pytest options for the grammar-fixture knobs** (folded into #2): all
   former `GRAMMAR_*` module-level constants are now `--grammar-*` CLI
   options registered via `pytest_addoption` in `tests/grammar/conftest.py`.
   Defaults are the values the PR thread calibrated against to produce
   unambiguous distributional + strict grammar-validity signal. Run
   `pytest tests/grammar/test_cli.py --grammar-pretrain-epochs=800
   --grammar-n-held-out=32` for debugging sweeps without editing code.

Follow-ups filed (not in this commit):
- meds-testing-helpers #40: consider upstreaming the grammar-synthetic-
  dataset generator pattern.

Verified: `pytest tests/grammar/test_in_process.py` all pass (4/4 in 20s);
`tests/grammar/_grammar.py` doctests pass; lint clean.
