+++ conftest.py
@@ -16,16 +16,6 @@ from torch.utils.data import DataLoader
 
 from MEDS_EIC_AR.model.model import Model
 from MEDS_EIC_AR.training.module import MEICARModule
-from tests._grammar_meds import (
-    GRAMMAR_GENERATE_BATCH_SIZE,
-    GRAMMAR_MAX_SEQ_LEN,
-    GRAMMAR_MODEL_HEAD_DIM,
-    GRAMMAR_MODEL_HEADS,
-    GRAMMAR_N_TRAJECTORIES,
-    GRAMMAR_PRETRAIN_BATCH_SIZE,
-    GRAMMAR_PRETRAIN_EPOCHS,
-    GRAMMAR_ROLLING_MAX_NEW_TOKENS,
-)
 
 
 def run_and_check(cmd: list[str]) -> dict[str, str]:
@@ -259,153 +249,7 @@ def generated_trajectories_rolling(
     return output_dir
 
 
-# ---------------------------------------------------------------------------
-# Grammar-dataset fixtures (issue #105) — pretrain + generate on a synthetic
-# grammar dataset via the full CLI pipeline so we can validate generation
-# correctness end-to-end, not just plumbing.
-# ---------------------------------------------------------------------------
-
-
-# Training-hyperparameter constants live in ``tests/_grammar_meds`` so pretrain and generate
-# fixtures + CLI test assertions all agree on them. See issue #105.
-
-
-@pytest.fixture(scope="session")
-def grammar_raw_meds(tmp_path_factory: pytest.TempPathFactory) -> tuple[Path, Path]:
-    """Build a raw-MEDS directory from the synthetic grammar + a task-labels directory.
-
-    Returns a tuple ``(meds_root, task_labels_dir)`` that matches the shape of
-    ``preprocessed_dataset_with_task``: the MEDS root has ``data/`` and ``metadata/`` subdirs
-    ready for ``MEICAR_process_data``, and the task-labels dir has per-split parquets keyed by
-    subject with a single prediction time per subject.
-    """
-    from tests._grammar_meds import build_grammar_meds_dataset
-
-    root = tmp_path_factory.mktemp("grammar_raw_meds")
-    meds_dir = root / "meds"
-    task_dir = root / "task_labels"
-    build_grammar_meds_dataset(meds_dir, task_labels_dir=task_dir)
-    return meds_dir, task_dir
-
-
-@pytest.fixture(scope="session")
-def grammar_preprocessed(
-    grammar_raw_meds: tuple[Path, Path], tmp_path_factory: pytest.TempPathFactory
-) -> Path:
-    """Runs ``MEICAR_process_data`` on the grammar raw MEDS dir and returns the tensorized dir."""
-    meds_dir, _ = grammar_raw_meds
-    return run_process_data(meds_dir, tmp_path_factory.mktemp("grammar_preprocessed"))
-
-
-@pytest.fixture(scope="session")
-def grammar_pretrained(grammar_preprocessed: Path, tmp_path_factory: pytest.TempPathFactory) -> Path:
-    """Runs ``MEICAR_pretrain`` on the grammar preprocessed dataset.
-
-    Uses a substantially bigger model than the default demo preset — see the ``GRAMMAR_*``
-    constants in ``tests/_grammar_meds``: hidden_size is ``GRAMMAR_MODEL_HEADS x
-    GRAMMAR_MODEL_HEAD_DIM``, training runs for ``GRAMMAR_PRETRAIN_EPOCHS`` with
-    ``GRAMMAR_PRETRAIN_BATCH_SIZE``, and early stopping is effectively disabled
-    (``patience=100000``). The budget is sized to dominate test cost on purpose: this fixture
-    underpins the strict greedy-correctness assertion in
-    ``test_grammar_cli_greedy_output_is_fully_grammar_valid``, which requires enough training
-    depth to reach 100% held-out grammar validity under greedy decoding — see the rationale on
-    ``GRAMMAR_PRETRAIN_EPOCHS`` for the forward-pass-count comparison with
-    ``test_pattern_generation.py``'s in-process budget.
-
-    Expected runtime: several minutes on a laptop CPU. This is the bulk of the grammar
-    integration test cost; everything downstream is cheap by comparison. Skip locally in favor
-    of doctests during iteration.
-    """
-    output_dir = tmp_path_factory.mktemp("grammar_pretrained")
-    run_and_check(
-        [
-            "MEICAR_pretrain",
-            "--config-name=_demo_pretrain",
-            f"output_dir={output_dir!s}",
-            f"datamodule.config.tensorized_cohort_dir={grammar_preprocessed!s}",
-            f"datamodule.batch_size={GRAMMAR_PRETRAIN_BATCH_SIZE}",
-            f"trainer.max_epochs={GRAMMAR_PRETRAIN_EPOCHS}",
-            "trainer.overfit_batches=0",
-            "trainer.callbacks.early_stopping.patience=100000",
-            f"max_seq_len={GRAMMAR_MAX_SEQ_LEN}",
-            f"lightning_module.model.gpt_kwargs.num_attention_heads={GRAMMAR_MODEL_HEADS}",
-            f"lightning_module.model.gpt_kwargs.attention_head_dim={GRAMMAR_MODEL_HEAD_DIM}",
-        ]
-    )
-    return output_dir
-
-
-def _run_grammar_generate(
-    output_dir: Path,
-    grammar_pretrained: Path,
-    grammar_preprocessed: Path,
-    task_dir: Path,
-    *,
-    do_sample: bool,
-) -> None:
-    """Invoke ``MEICAR_generate_trajectories`` with the shared grammar fixture config.
-
-    Centralized here so the sampling and greedy fixtures only differ by the ``do_sample`` flag
-    and can't drift apart on other knobs (batch size, rolling budget, N trajectories). All other
-    parameters come from the shared constants in :mod:`tests._grammar_meds`.
-    """
-    run_and_check(
-        [
-            "MEICAR_generate_trajectories",
-            "--config-name=_demo_generate_trajectories",
-            f"output_dir={output_dir!s}",
-            f"model_initialization_dir={grammar_pretrained!s}",
-            f"datamodule.config.tensorized_cohort_dir={grammar_preprocessed!s}",
-            f"datamodule.config.task_labels_dir={task_dir!s}",
-            f"datamodule.batch_size={GRAMMAR_GENERATE_BATCH_SIZE}",
-            "trainer=demo",
-            f"inference.N_trajectories_per_task_sample={GRAMMAR_N_TRAJECTORIES}",
-            f"inference.do_sample={str(do_sample).lower()}",
-            f"rolling_generation.max_new_tokens={GRAMMAR_ROLLING_MAX_NEW_TOKENS}",
-        ]
-    )
-
-
-@pytest.fixture(scope="session")
-def grammar_generated_trajectories(
-    grammar_pretrained: Path,
-    grammar_preprocessed: Path,
-    grammar_raw_meds: tuple[Path, Path],
-    tmp_path_factory: pytest.TempPathFactory,
-) -> Path:
-    """Runs ``MEICAR_generate_trajectories`` on the grammar-pretrained model with sampling.
-
-    Uses a larger ``N_trajectories_per_task_sample`` (vs. the demo default of 2) and a
-    ``batch_size`` that is deliberately *not* a multiple of N, so the interleaving / demux
-    code path from #89 is stressed in its cross-sample-group configuration. Rolling generation
-    is enabled with a budget above ``max_seq_len`` so we also exercise the sliding-window path
-    end-to-end.
-    """
-    _, task_dir = grammar_raw_meds
-    output_dir = tmp_path_factory.mktemp("grammar_generated_trajectories")
-    _run_grammar_generate(output_dir, grammar_pretrained, grammar_preprocessed, task_dir, do_sample=True)
-    return output_dir
-
-
-@pytest.fixture(scope="session")
-def grammar_generated_trajectories_greedy(
-    grammar_pretrained: Path,
-    grammar_preprocessed: Path,
-    grammar_raw_meds: tuple[Path, Path],
-    tmp_path_factory: pytest.TempPathFactory,
-) -> Path:
-    """Same as :func:`grammar_generated_trajectories` but with ``do_sample=False`` (greedy argmax).
-
-    Greedy decoding gives fully deterministic outputs given the trained weights and prompt, so
-    this fixture underpins the strict grammar-correctness assertion in
-    ``test_grammar_cli_greedy_output_is_fully_grammar_valid``: if the model has learned the
-    grammar, its argmax continuation should be 100% grammar-valid, matching the in-process
-    :mod:`tests.test_pattern_generation` tests' strictness but through the full CLI pipeline.
-    """
-    _, task_dir = grammar_raw_meds
-    output_dir = tmp_path_factory.mktemp("grammar_generated_trajectories_greedy")
-    _run_grammar_generate(output_dir, grammar_pretrained, grammar_preprocessed, task_dir, do_sample=False)
-    return output_dir
+# Grammar-dataset fixtures (issue #105) live in ``tests/grammar/conftest.py``.
 
 
 @contextmanager
